chore: refactor birdy test const/utils
