+++ src/components/signup/BirdyTestForm.tsx
@@ -3,7 +3,7 @@
 import { useRouter } from "next/navigation";
 import { useBirdyTestStore } from "@/store/useBirdyTestStore";
 import ProgressBar from "../ui/ProgressBar";
-import { questions } from "@/constants/birdyTestQuestions";
+import { questions } from "@/constants/birdyTest";
 import clsx from "clsx";
 
 export const ProgressLabel = ({
